requirements.txt
configparser==7.0.0
dbus-python==1.3.2
PyGObject==3.48.2
nuitka
pyinstaller